Fix and merge
