+++ dark/views/tournament/team/change_name.py
@@ -2,9 +2,9 @@ from django.contrib import messages
 from django.views.generic import UpdateView
 from django.urls import reverse
 
-from dark.common.views import TournamentEditableMixin
 from dark.models.tournament.team import Team
 from dark.forms.tournament.team import ChangeTeamNameForm
+from dark.views.tournament.mixins import TournamentEditableMixin
 
 
 class ChangeTeamNameView(TournamentEditableMixin, UpdateView):
