+++ dark/views/tournament/team/member/join.py
@@ -3,8 +3,8 @@ from django.http import HttpResponseRedirect
 from django.urls import reverse
 from django.views.generic.base import View
 
-from dark.common.views import TournamentEditableMixin
 from dark.models.tournament.team import Team, TeamRole, TeamMember
+from dark.views.tournament.mixins import TournamentEditableMixin
 
 
 class JoinTeamMemberView(TournamentEditableMixin, View):
